Moves PromptTemplate & PromptLocation to respecive templtes and azureutil modules
