+++ cli/azd/test/mocks/console/mock_console.go
@@ -6,13 +6,10 @@ import (
 	"io"
 
 	"github.com/azure/azure-dev/cli/azd/pkg/input"
-	"github.com/azure/azure-dev/cli/azd/pkg/templates"
 )
 
 // A predicate function definition for registering expressions
 type WhenPredicate func(options input.ConsoleOptions) bool
-
-var truePredicateFn WhenPredicate = func(_ input.ConsoleOptions) bool { return true }
 
 // A mock implementation of the input.Console interface
 type MockConsole struct {
@@ -58,20 +55,6 @@ func (c *MockConsole) Select(ctx context.Context, options input.ConsoleOptions) 
 	c.log = append(c.log, options.Message)
 	value, err := c.respond("Select", options)
 	return value.(int), err
-}
-
-// Writes a location selection choice to the console for the user to choose
-func (c *MockConsole) PromptLocation(ctx context.Context, message string) (string, error) {
-	c.log = append(c.log, message)
-	value, err := c.respond("PromptLocation", input.ConsoleOptions{})
-	return value.(string), err
-}
-
-// Writes a template selection choice to the console for the user to choose
-func (c *MockConsole) PromptTemplate(ctx context.Context, message string) (templates.Template, error) {
-	c.log = append(c.log, message)
-	value, err := c.respond("PromptTemplate", input.ConsoleOptions{})
-	return value.(templates.Template), err
 }
 
 // Writes messages to the underlying writer
@@ -130,29 +113,6 @@ func (c *MockConsole) WhenSelect(predicate WhenPredicate) *MockConsoleExpression
 
 	c.expressions = append(c.expressions, &expr)
 	return &expr
-}
-
-// Registers a location prompt selection for mocking in unit tests
-func (c *MockConsole) WhenPromptLocation() *MockConsoleExpression {
-	expr := MockConsoleExpression{
-		command:     "PromptLocation",
-		console:     c,
-		predicateFn: truePredicateFn,
-	}
-
-	c.expressions = append(c.expressions, &expr)
-	return &expr
-}
-
-// Registers a template prompt selection for mocking in unit tests
-func (c *MockConsole) WhenPromptTemplate() *MockConsoleExpression {
-	expression := &MockConsoleExpression{
-		command:     "PromptTemplate",
-		console:     c,
-		predicateFn: truePredicateFn,
-	}
-
-	return expression
 }
 
 // MockConsoleExpression is an expression with options response or error
